remove abstract model and modelwrapper
- models must implement SectorModel to be run
- SectorModel exposes common interface for smif
- SosModel contains and coordinates sector models
